# api/api.yaml
openapi: 3.0.0
info:
  title: HackerNews API
  version: "1.0.0"
servers:
  - url: 'https://asw-hackernwes-g11.herokuapp.com'
    description: Heroku server
...
paths:
  /tweets:
    get:
      summary: Llista tots les publicacions
      tags: 
        - Publicacions
      parameters:
      - name: orderby
        required: false
        in: query
        description: Ordre en què es volen llistar les publicacions
        schema:
          type: string
      - name: user_id
        required: false
        in: query
        description: Id de l'usuari
        schema:
          type: integer
          minimum: 1  
      - name: voted
        required: false
        in: query
        description: Indica si es volen les contribucions que l'usuari ha votat
        schema:
          type: boolean
      responses:
        200:
          description: Retorna totes les publicacions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    author:
                      type: string
                      example: user1
                    content:
                      type: string
                      example: ""
                    created_at:
                      type: string
                      example: "2021-05-08T00:52:46.169Z"
                    url:
                      type: string
                      example: www.fib.upc.edu
                    title:
                      type: string
                      example: La FIB
                  required:
                    - id
                    - author
                    - content
                    - created_at
                    - url
                    - title
      security:
        - ApiKeyAuth: []
        
    post:
      summary: Creació d'una publicació
      tags: 
        - Publicacions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: La FIB
                url:
                  type: string
                  example: www.fib.upc.edu
                content:
                  type: string
                  example: Aquesta és la meva facultat
              required:
                - title
      responses:
        201:
          description: Publicació creada correctament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tweet'
        401:
          description: La crida no conté una api_key vàlida
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized, you don't have a valid api_key
        400:
          description: "Error - Aquesta URL ja existeix, el titol està buit o s'ha omplert el content i la url a la vegada"
          content:
            application/json:
              schema:
                type: string
                example: The URL already exists
      security:
        - ApiKeyAuth: []
  /tweets/{id}:
    get:
      summary: Permet veure, votar o desvotar una publicació
      tags: 
        - Publicacions
      parameters: 
        - name: id
          in: path
          required: true
          description: Id de la publicació
          schema:
            type: integer
            minimum: 1
            example: 1
        - name: vote
          in: query
          description: La intenció de votar o desvotar una publicació
          schema:
            type: boolean
      responses:
        200: 
          description: "No hi ha hagut cap error, es mostra la contribució o bé s'indica que s'ha votat o desvotat correctament"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Tweet'
        400:
          description: Error - Ja has votat/desvotat aquesta contribució
          content:
            application/json:
              schema:
                type: string
                example: You have alredy voted this contribution
        403: 
          description: Error - No pots votar una contribució que és teva
          content:
            application/json:
              schema:
                type: string
                example: You can't vote your contribution
        404:
          description: 'La contribució no existeix'
      security:
        - ApiKeyAuth: []
    delete:
      summary: Elimina una publicació
      parameters:
        - name: id
          required: true
          in: path
          description: Id de la publicació
          schema:
            type: integer
            minimum: 1
            example: 1
      security:
        - ApiKeyAuth: []
      tags:
        - Publicacions
      responses:
        200:
          description: Publicació eliminada correctament
          content:
            application/json:
              schema:
                type: string
                example: Contribution successfully deleted
        401:
          description: La crida no conté una api_key vàlida
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized, you don't have a valid api_key
        403:
          description: "L'usuari no té permisos per eliminar la publicació"
          content:
            application/json:
              schema:
                type: string
                example: "The user does not have permissions to delete this contribution"
        404:
          description: 'La publicació no existeix'
    put:
      summary: Editar una publicació
      tags: 
        - Publicacions
      parameters:
        - name: id
          required: true
          in: path
          description: Id de la publicació
          schema:
            type: integer
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: La FIB
                url:
                  type: string
                  example: www.fib.upc.edu
                content:
                  type: string
                  example: Aquesta és la meva facultat
              required:
                - title
      responses:
        200:
          description: Publicació editada correctament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tweet'
        401:
          description: La crida no conté una api_key vàlida
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized, you don't have a valid api_key
        400:
          description: "Error - El títol està en blanc o s'ha omplert el camp de text quan la publicació no té el camp URL buit"
          content:
            application/json:
              schema:
                type: string
                example: The title is blank
        403:
          description: 'Error - No pots editar una publicació que no és teva'
          content:
            application/json:
              schema:
                type: string
                example: You can't edit a contribution that is not yours
        404:
          description: 'La publicació no existeix'
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
    Tweet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        author:
          type: string
          example: user1
        content:
          type: string
          example: ""
        created_at:
          type: string
          example: "2021-05-08T00:52:46.169Z"
        url:
          type: string
          example: www.fib.upc.edu
        title:
          type: string
          example: La FIB
      required:
        - id
        - author
        - content
        - created_at
        - url
        - title
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY
      in: header